update 'create,delete,update' comment to use auth

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -134,13 +134,15 @@ const Mutation = {
     }, info)
   }, //updatePost
 
-  createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -152,7 +154,17 @@ const Mutation = {
     }, info)
   }, // createComment
 
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    }) // commentExists
+
+    if (!commentExists) throw new Error('Unable to delete comment')
+
     return prisma.mutation.deleteComment({
       where: {
         id: args.id
@@ -160,7 +172,17 @@ const Mutation = {
     }, info)
   }, // deleteComment
 
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    }) // commentExists
+
+    if (!commentExists) throw new Error('Unable to update comment')
+
     return prisma.mutation.updateComment({
       where: {
         id: args.id
